Verbesserungen Fontend + Bugfixes
